Initial release
This was a library I ran within a console application. I wrote it to automate a large and tedious workload that had been assigned to me. As I wrote what I needed I also gave some thought to how it might evolve and so parts of it are stubbed out and incomplete.